+++ packages/components/src/composite/index.tsx
@@ -33,10 +33,11 @@ import type {
 
 /**
  * Creates a composite store.
- * @param props
- * @see https://ariakit.org/reference/use-composite-store
+ *
  * @example
  * ```jsx
+ * import { Composite, useCompositeStore } from '@wordpress/components';
+ *
  * const store = useCompositeStore();
  * <Composite store={store}>
  *   <Composite.Item>Item</Composite.Item>
@@ -45,8 +46,24 @@ import type {
  * </Composite>
  * ```
  */
-export function useCompositeStore( props: CompositeStoreProps ) {
-	return Ariakit.useCompositeStore( props );
+export function useCompositeStore( {
+	focusLoop = false,
+	focusWrap = false,
+	focusShift = false,
+	virtualFocus = false,
+	orientation = 'both',
+	rtl = false,
+	...props
+}: CompositeStoreProps = {} ) {
+	return Ariakit.useCompositeStore( {
+		focusLoop,
+		focusWrap,
+		focusShift,
+		virtualFocus,
+		orientation,
+		rtl,
+		...props,
+	} );
 }
 
 const Group = forwardRef<
@@ -82,10 +99,14 @@ const Row = forwardRef<
 Row.displayName = 'Composite.Row';
 
 /**
- * Renders a composite widget.
- * @see https://ariakit.org/reference/composite
+ * Renders a widget based on the WAI-ARIA [`composite`](https://w3c.github.io/aria/#composite)
+ * role, which provides a single tab stop on the page and arrow key navigation
+ * through the focusable descendants.
+ *
  * @example
  * ```jsx
+ * import { Composite, useCompositeStore } from '@wordpress/components';
+ *
  * const store = useCompositeStore();
  * <Composite store={store}>
  *   <Composite.Item>Item 1</Composite.Item>
@@ -104,9 +125,11 @@ export const Composite = Object.assign(
 		displayName: 'Composite',
 		/**
 		 * Renders a group element for composite items.
-		 * @see https://ariakit.org/reference/composite-group
+		 *
 		 * @example
 		 * ```jsx
+		 * import { Composite, useCompositeStore } from '@wordpress/components';
+		 *
 		 * const store = useCompositeStore();
 		 * <Composite store={store}>
 		 *   <Composite.Group>
@@ -122,9 +145,11 @@ export const Composite = Object.assign(
 		 * Renders a label in a composite group. This component must be wrapped with
 		 * `Composite.Group` so the `aria-labelledby` prop is properly set on the
 		 * composite group element.
-		 * @see https://ariakit.org/reference/composite-group-label
+		 *
 		 * @example
 		 * ```jsx
+		 * import { Composite, useCompositeStore } from '@wordpress/components';
+		 *
 		 * const store = useCompositeStore();
 		 * <Composite store={store}>
 		 *   <Composite.Group>
@@ -138,9 +163,11 @@ export const Composite = Object.assign(
 		GroupLabel,
 		/**
 		 * Renders a composite item.
-		 * @see https://ariakit.org/reference/composite-item
+		 *
 		 * @example
 		 * ```jsx
+		 * import { Composite, useCompositeStore } from '@wordpress/components';
+		 *
 		 * const store = useCompositeStore();
 		 * <Composite store={store}>
 		 *   <Composite.Item>Item 1</Composite.Item>
@@ -154,9 +181,11 @@ export const Composite = Object.assign(
 		 * Renders a composite row. Wrapping `Composite.Item` elements within
 		 * `Composite.Row` will create a two-dimensional composite widget, such as a
 		 * grid.
-		 * @see https://ariakit.org/reference/composite-row
+		 *
 		 * @example
 		 * ```jsx
+		 * import { Composite, useCompositeStore } from '@wordpress/components';
+		 *
 		 * const store = useCompositeStore();
 		 * <Composite store={store}>
 		 *   <Composite.Row>